Updated Token helper/manager
